requirements.txt
# Core ML dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
joblib>=1.1.0

# Visualization (optional, for notebook)
matplotlib>=3.5.0
plotly>=5.0.0

# Hardware monitoring
pythonnet>=3.0.0

# GUI and System Tray
pystray>=0.19.0
Pillow>=10.0.0

# Notifications
plyer>=2.1.0

# Configuration
PyYAML>=6.0
